+++ script/tool/test/native_test_command_test.dart
@@ -736,6 +736,146 @@ void main() {
       });
     });
 
+    group('Linux', () {
+      test('runs unit tests', () async {
+        const String testBinaryRelativePath =
+            'build/linux/foo/debug/bar/plugin_test';
+        final Directory pluginDirectory =
+            createFakePlugin('plugin', packagesDir, extraFiles: <String>[
+          'example/$testBinaryRelativePath'
+        ], platformSupport: <String, PlatformDetails>{
+          kPlatformLinux: const PlatformDetails(PlatformSupport.inline),
+        });
+
+        final File testBinary = childFileWithSubcomponents(pluginDirectory,
+            <String>['example', ...testBinaryRelativePath.split('/')]);
+
+        final List<String> output = await runCapturingPrint(runner, <String>[
+          'native-test',
+          '--linux',
+          '--no-integration',
+        ]);
+
+        expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('Running plugin_test...'),
+            contains('No issues found!'),
+          ]),
+        );
+
+        expect(
+            processRunner.recordedCalls,
+            orderedEquals(<ProcessCall>[
+              ProcessCall(testBinary.path, const <String>[], null),
+            ]));
+      });
+
+      test('only runs debug unit tests', () async {
+        const String debugTestBinaryRelativePath =
+            'build/linux/foo/debug/bar/plugin_test';
+        const String releaseTestBinaryRelativePath =
+            'build/linux/foo/release/bar/plugin_test';
+        final Directory pluginDirectory =
+            createFakePlugin('plugin', packagesDir, extraFiles: <String>[
+          'example/$debugTestBinaryRelativePath',
+          'example/$releaseTestBinaryRelativePath'
+        ], platformSupport: <String, PlatformDetails>{
+          kPlatformLinux: const PlatformDetails(PlatformSupport.inline),
+        });
+
+        final File debugTestBinary = childFileWithSubcomponents(pluginDirectory,
+            <String>['example', ...debugTestBinaryRelativePath.split('/')]);
+
+        final List<String> output = await runCapturingPrint(runner, <String>[
+          'native-test',
+          '--linux',
+          '--no-integration',
+        ]);
+
+        expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('Running plugin_test...'),
+            contains('No issues found!'),
+          ]),
+        );
+
+        // Only the debug version should be run.
+        expect(
+            processRunner.recordedCalls,
+            orderedEquals(<ProcessCall>[
+              ProcessCall(debugTestBinary.path, const <String>[], null),
+            ]));
+      });
+
+      test('fails if there are no unit tests', () async {
+        createFakePlugin('plugin', packagesDir,
+            platformSupport: <String, PlatformDetails>{
+              kPlatformLinux: const PlatformDetails(PlatformSupport.inline),
+            });
+
+        Error? commandError;
+        final List<String> output = await runCapturingPrint(runner, <String>[
+          'native-test',
+          '--linux',
+          '--no-integration',
+        ], errorHandler: (Error e) {
+          commandError = e;
+        });
+
+        expect(commandError, isA<ToolExit>());
+        expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('No test binaries found.'),
+          ]),
+        );
+
+        expect(processRunner.recordedCalls, orderedEquals(<ProcessCall>[]));
+      });
+
+      test('fails if a unit test fails', () async {
+        const String testBinaryRelativePath =
+            'build/linux/foo/debug/bar/plugin_test';
+        final Directory pluginDirectory =
+            createFakePlugin('plugin', packagesDir, extraFiles: <String>[
+          'example/$testBinaryRelativePath'
+        ], platformSupport: <String, PlatformDetails>{
+          kPlatformLinux: const PlatformDetails(PlatformSupport.inline),
+        });
+
+        final File testBinary = childFileWithSubcomponents(pluginDirectory,
+            <String>['example', ...testBinaryRelativePath.split('/')]);
+
+        processRunner.mockProcessesForExecutable[testBinary.path] =
+            <io.Process>[MockProcess(exitCode: 1)];
+
+        Error? commandError;
+        final List<String> output = await runCapturingPrint(runner, <String>[
+          'native-test',
+          '--linux',
+          '--no-integration',
+        ], errorHandler: (Error e) {
+          commandError = e;
+        });
+
+        expect(commandError, isA<ToolExit>());
+        expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('Running plugin_test...'),
+          ]),
+        );
+
+        expect(
+            processRunner.recordedCalls,
+            orderedEquals(<ProcessCall>[
+              ProcessCall(testBinary.path, const <String>[], null),
+            ]));
+      });
+    });
+
     // Tests behaviors of implementation that is shared between iOS and macOS.
     group('iOS/macOS', () {
       test('fails if xcrun fails', () async {