+++ script/tool/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+- `native-test` now supports `--linux` for unit tests.
+
 ## 0.6.0
 
 - Added Android native integration test support to `native-test`.